requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
Werkzeug==2.3.6
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6